+++ backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/konan/llvm/RTTIGenerator.kt
@@ -16,6 +16,7 @@ import org.jetbrains.kotlin.ir.declarations.IrProperty
 import org.jetbrains.kotlin.ir.types.*
 import org.jetbrains.kotlin.ir.util.isAnnotationClass
 import org.jetbrains.kotlin.ir.util.isInterface
+import org.jetbrains.kotlin.konan.KonanAbiVersion
 import org.jetbrains.kotlin.name.FqName
 
 internal class RTTIGenerator(override val context: Context) : ContextUtils {
@@ -59,18 +60,17 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
                 result = result or TF_ACYCLIC
             }
         }
+        if (classDescriptor.isInterface)
+            result = result or TF_INTERFACE
         return result
     }
-
-    private inner class FieldTableRecord(val nameSignature: LocalHash, fieldOffset: Int) :
-            Struct(runtime.fieldTableRecordType, nameSignature, Int32(fieldOffset))
 
     inner class MethodTableRecord(val nameSignature: LocalHash, methodEntryPoint: ConstPointer?) :
             Struct(runtime.methodTableRecordType, nameSignature, methodEntryPoint)
 
     private inner class TypeInfo(
             selfPtr: ConstPointer,
-            name: ConstValue,
+            extendedInfo: ConstPointer,
             size: Int,
             superType: ConstValue,
             objOffsets: ConstValue,
@@ -79,12 +79,9 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
             interfacesCount: Int,
             methods: ConstValue,
             methodsCount: Int,
-            fields: ConstValue,
-            fieldsCount: Int,
             packageName: String?,
             relativeName: String?,
             flags: Int,
-            extendedInfo: ConstPointer,
             writableTypeInfo: ConstPointer?) :
 
             Struct(
@@ -92,7 +89,10 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
 
                     selfPtr,
 
-                    name,
+                    extendedInfo,
+
+                    Int32(KonanAbiVersion.CURRENT.version),
+
                     Int32(size),
 
                     superType,
@@ -106,15 +106,10 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
                     methods,
                     Int32(methodsCount),
 
-                    fields,
-                    Int32(fieldsCount),
-
                     kotlinStringLiteral(packageName),
                     kotlinStringLiteral(relativeName),
 
                     Int32(flags),
-
-                    extendedInfo,
 
                     *listOfNotNull(writableTypeInfo).toTypedArray()
             )
@@ -180,8 +175,6 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
 
         val bodyType = llvmDeclarations.bodyType
 
-        val name = className.globalHash
-
         val size = getInstanceSize(bodyType, className)
 
         val superTypeOrAny = classDesc.getSuperClassNotAny() ?: context.ir.symbols.any.owner
@@ -192,7 +185,6 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
         val interfacesPtr = staticData.placeGlobalConstArray("kintf:$className",
                 pointerType(runtime.typeInfoType), interfaces)
 
-        // TODO: reuse offsets obtained for 'fields' below
         val objOffsets = getStructElements(bodyType).mapIndexedNotNull { index, type ->
             if (isObjectType(type)) {
                 LLVMOffsetOfElement(llvmTargetData, bodyType, index)
@@ -210,16 +202,6 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
             objOffsets.size
         }
 
-        val fields = llvmDeclarations.fields.mapIndexed { index, field ->
-            // Note: using FQ name because a class may have multiple fields with the same name due to property overriding
-            val nameSignature = field.fqNameSafe.localHash // FIXME: add signature
-            val fieldOffset = LLVMOffsetOfElement(llvmTargetData, bodyType, index)
-            FieldTableRecord(nameSignature, fieldOffset.toInt())
-        }.sortedBy { it.nameSignature.value }
-
-        val fieldsPtr = staticData.placeGlobalConstArray("kfields:$className",
-                runtime.fieldTableRecordType, fields)
-
         val methods = if (classDesc.isAbstract()) {
             emptyList()
         } else {
@@ -233,17 +215,15 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
         val typeInfoGlobal = llvmDeclarations.typeInfoGlobal
         val typeInfo = TypeInfo(
                 classDesc.typeInfoPtr,
-                name,
+                makeExtendedInfo(classDesc),
                 size,
                 superType,
                 objOffsetsPtr, objOffsetsCount,
                 interfacesPtr, interfaces.size,
                 methodsPtr, methods.size,
-                fieldsPtr, if (classDesc.isInterface) -1 else fields.size,
                 reflectionInfo.packageName,
                 reflectionInfo.relativeName,
                 flagsFromClass(classDesc),
-                makeExtendedInfo(classDesc),
                 llvmDeclarations.writableTypeInfoGlobal?.pointer
         )
 
@@ -309,9 +289,9 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
                     NullPointer(int32Type), NullPointer(int8Type), NullPointer(kInt8Ptr))
         } else {
             data class FieldRecord(val offset: Int, val type: Int, val name: String)
-            val fields = getStructElements(bodyType).mapIndexedNotNull { index, type ->
+            val fields = getStructElements(bodyType).drop(1).mapIndexedNotNull { index, type ->
                 FieldRecord(
-                        LLVMOffsetOfElement(llvmTargetData, bodyType, index).toInt(),
+                        LLVMOffsetOfElement(llvmTargetData, bodyType, index + 1).toInt(),
                         mapRuntimeType(type),
                         llvmDeclarations.fields[index].name.asString())
             }
@@ -324,6 +304,7 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
             Struct(runtime.extendedTypeInfoType, Int32(fields.size), offsetsPtr, typesPtr, namesPtr)
         }
         val result = staticData.placeGlobal("", value)
+        result.setConstant(true)
         return result.pointer
     }
 
@@ -335,9 +316,7 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
     ): ConstPointer {
         assert(descriptor.isInterface)
 
-        val name = "".globalHash
-
-        val size = 0
+        val size = LLVMStoreSizeOfType(llvmTargetData, kObjHeader).toInt()
 
         val superClass = context.ir.symbols.any.owner
 
@@ -349,8 +328,6 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
         assert(superClass.declarations.all { it !is IrProperty && it !is IrField })
         val objOffsetsPtr = NullPointer(int32Type)
         val objOffsetsCount = 0
-        val fieldsPtr = NullPointer(runtime.fieldTableRecordType)
-        val fieldsCount = 0
 
         val methods = (methodTableRecords(superClass) + methodImpls.map { (method, impl) ->
             assert(method.containingDeclaration == descriptor)
@@ -377,17 +354,15 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
         val result = typeInfoWithVtableGlobal.pointer.getElementPtr(0)
         val typeInfoWithVtable = Struct(TypeInfo(
                 selfPtr = result,
-                name = name,
+                extendedInfo = NullPointer(runtime.extendedTypeInfoType),
                 size = size,
                 superType = superClass.typeInfoPtr,
                 objOffsets = objOffsetsPtr, objOffsetsCount = objOffsetsCount,
                 interfaces = interfacesPtr, interfacesCount = interfaces.size,
                 methods = methodsPtr, methodsCount = methods.size,
-                fields = fieldsPtr, fieldsCount = fieldsCount,
                 packageName = reflectionInfo.packageName,
                 relativeName = reflectionInfo.relativeName,
                 flags = flagsFromClass(descriptor) or (if (immutable) TF_IMMUTABLE else 0),
-                extendedInfo = NullPointer(runtime.extendedTypeInfoType),
                 writableTypeInfo = writableTypeInfo
               ), vtable)
 
@@ -417,5 +392,7 @@ internal class RTTIGenerator(override val context: Context) : ContextUtils {
     }
 }
 
+// Keep in sync with Konan_TypeFlags in TypeInfo.h.
 private const val TF_IMMUTABLE = 1
 private const val TF_ACYCLIC   = 2
+private const val TF_INTERFACE = 4