+++ runtime/src/main/kotlin/kotlin/native/internal/RuntimeUtils.kt
@@ -32,7 +32,6 @@ fun ThrowInvalidReceiverTypeException(klass: KClass<*>): Nothing {
     throw RuntimeException("Unexpected receiver type: " + (klass.qualifiedName ?: "noname"))
 }
 
-@ExportForCppRuntime
 internal fun ThrowArithmeticException() : Nothing {
     throw ArithmeticException()
 }
@@ -58,6 +57,11 @@ fun ThrowUninitializedPropertyAccessException(): Nothing {
 @ExportForCppRuntime
 internal fun ThrowIllegalArgumentException() : Nothing {
     throw IllegalArgumentException()
+}
+
+@ExportForCppRuntime
+internal fun ThrowIllegalStateException() : Nothing {
+    throw IllegalStateException()
 }
 
 @ExportForCppRuntime